test: address zkochan + Copilot review feedback on #11710
- Move executor + fs::file_mode unit tests from inline modules
  to the project's standard `src/<parent>/tests.rs` layout. The
  convention is documented in `CODE_STYLE_GUIDE.md` under "Unit
  test file layout"; zkochan flagged the violation in the
  executor file.

- Gate `pacquet-executor::tests` on `cfg(all(test, unix))` at
  the declaration site so the `use super::execute_shell` import
  isn't dead on Windows. Copilot flagged that
  `clippy --tests --deny warnings` would fail there.

- Drop `execute_shell_propagates_nonzero_exit_as_ok`. Copilot
  flagged that it regression-pinned a behavior that *should*
  change: pnpm/npm `run` exits with the script's status, but
  pacquet's wrapper currently returns `Ok(())` regardless. A
  fix belongs in its own PR; until then we should not add tests
  that block that fix.

- Quote the temp-path redirect targets in the `cli/tests/run.rs`
  shell fixtures so a tempdir path containing a space
  (`/var/folders/...` on macOS) doesn't split the `touch` /
  redirect argument. Copilot flagged the unquoted paths.

The `fs::file_mode::make_file_executable_sets_exec_bits` test
keeps `#[cfg(unix)]` at the test site (and moves its
`make_file_executable` import inside) rather than gating the
whole `tests` module — the other two tests
(`exec_constants_pin_pnpm_layout`,
`is_executable_matches_any_exec_bit`) are platform-neutral and
should still run on Windows.

Copilot's concern about `detect_node_version`'s `expect` (in
`engine_name.rs`) is intentionally not addressed: KSXGitHub
asked for the hard-fail in 5d0987c on grounds that `node` is a
documented prerequisite of the test suite, and the human
reviewer's call wins over the bot's.
